+++ filters/olefilters/winword97/section.cc
@@ -20,8 +20,8 @@
 #include <section.h>
 
 Section::Section(const WinWordDoc * const parent, const unsigned char * const mainData,
-                 const FIB * const fib, const QArray<long> &paragMarks,
-                 const QArray<long> &cellMarks, const QArray<long> &rowMarks) :
+                 const FIB * const fib, const QArray<int> &paragMarks,
+                 const QArray<int> &cellMarks, const QArray<int> &rowMarks) :
                  m_parent(parent), m_mainData(mainData), m_fib(fib),
                  m_paragMarks(paragMarks), m_cellMarks(cellMarks),
                  m_rowMarks(rowMarks) {
